fixed norm and some bugs - sigs in childs , exit codes of sigs
