+++ executer/utils/sigs.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   sigs.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: aamoussa <aamoussa@student.42.fr>          +#+  +:+       +#+        */
+/*   By: zlafou <zlafou@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/18 00:22:15 by zlafou            #+#    #+#             */
-/*   Updated: 2023/01/26 18:52:07 by aamoussa         ###   ########.fr       */
+/*   Updated: 2023/01/29 20:16:17 by zlafou           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,12 @@
 
 void	sig_int(int signal)
 {
+	char	*s;
+
 	if (signal == SIGINT)
 	{
+		s = ps1();
 		g_gb.exit_statut = 1;
-		char *s = ps1();
 		printf("\33[2K\r> %s\n%s", rl_line_buffer, s);
 		rl_replace_line("", 0);
 		free(s);
@@ -25,7 +27,13 @@ void	sig_int(int signal)
 }
 
 void	signals(void)
-{    
+{
 	signal(SIGINT, sig_int);
 	signal(SIGQUIT, SIG_IGN);
+}
+
+void	ch_signals(void)
+{
+	signal(SIGINT, SIG_DFL);
+	signal(SIGQUIT, SIG_DFL);
 }
