+++ executer/utils/strings.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   strings.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: aamoussa <aamoussa@student.42.fr>          +#+  +:+       +#+        */
+/*   By: zlafou <zlafou@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/17 00:47:24 by zlafou            #+#    #+#             */
-/*   Updated: 2023/01/26 18:52:14 by aamoussa         ###   ########.fr       */
+/*   Updated: 2023/01/29 20:15:43 by zlafou           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
