+++ minishell.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   minishell.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: aamoussa <aamoussa@student.42.fr>          +#+  +:+       +#+        */
+/*   By: zlafou <zlafou@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/16 17:11:41 by aamoussa          #+#    #+#             */
-/*   Updated: 2023/01/26 19:03:35 by aamoussa         ###   ########.fr       */
+/*   Updated: 2023/01/29 20:28:02 by zlafou           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,7 +33,6 @@
 # define BLUE "\033[0;35m"
 # define PURPLE "\033[0;34m"
 # define RESET "\033[0;0m"
-
 
 typedef struct s_envn
 {
@@ -79,7 +78,11 @@ typedef struct s_pipecmd {
 	t_cmd	*right;
 }	t_pipecmd;
 
+void	leave(void);
 void	free_cmd(t_cmd *cmd);
+int		is_builtin(t_execcmd *x_cmd);
+void	exec_builtins(t_execcmd *x_cmd);
+void	n_pipe(t_cmd *cmd, int in, int out);
 t_cmd	*parser(char **ps, char *es, char **envp);
 void	rl_replace_line(const char *text, int clear_undo);
 int		ft_strcmp(const char *s1, const char *s2);
